@@ -1,81 +1,79 @@
-"use strict";
-
 module.exports = {
-	name: "greeter",
+  name: 'greeter',
 
-	/**
-	 * Service settings
-	 */
-	settings: {
+  /**
+   * Service settings
+   */
+  settings: {
 
-	},
+  },
 
-	/**
-	 * Service dependencies
-	 */
-	dependencies: [],	
+  /**
+   * Service dependencies
+   */
+  dependencies: [],
 
-	/**
-	 * Actions
-	 */
-	actions: {
+  /**
+   * Actions
+   */
+  actions: {
 
-		/**
-		 * Say a 'Hello'
-		 *
-		 * @returns
-		 */
-		hello() {
-			return "Hello Moleculer";
-		},
+    /**
+     * Say a 'Hello'
+     *
+     * @returns
+     */
+    hello () {
+      return 'Hello Moleculer'
+    },
 
-		/**
-		 * Welcome a username
-		 *
-		 * @param {String} name - User name
-		 */
-		welcome: {
-			params: {
-				name: "string"
-			},
-			handler(ctx) {
-				return `Welcome, ${ctx.params.name}`;
-			}
-		}
-	},
+    /**
+     * Welcome a username
+     *
+     * @param {String} name - User name
+     */
+    welcome: {
+      params: {
+        name: 'string'
+      },
+      handler (ctx) {
+        return `Welcome, ${ctx.params.name}`
+      }
+    }
+  },
 
-	/**
-	 * Events
-	 */
-	events: {
+  /**
+   * Events
+   */
+  events: {
 
-	},
+  },
 
-	/**
-	 * Methods
-	 */
-	methods: {
+  /**
+   * Methods
+   */
+  methods: {
 
-	},
+  },
 
-	/**
-	 * Service created lifecycle event handler
-	 */
-	created() {
+  /**
+   * Service created lifecycle event handler
+   */
+  created () {
 
-	},
+  },
 
-	/**
-	 * Service started lifecycle event handler
-	 */
-	started() {
+  /**
+   * Service started lifecycle event handler
+   */
+  started () {
 
-	},
+  },
 
-	/**
-	 * Service stopped lifecycle event handler
-	 */
-	stopped() {
+  /**
+   * Service stopped lifecycle event handler
+   */
+  stopped () {
 
-	}
-};
+  }
+}
